# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/missions/{mission}/result": {
            "post": {
                "tags": [
                    "Missions"
                ],
                "summary": "Добавить результат выполнения миссии пользователем",
                "description": "Позволяет пользователю добавить результат выполнения миссии (файл или текст)",
                "operationId": "5a62e00ff2f763ae96be05996dad50c3",
                "parameters": [
                    {
                        "name": "mission",
                        "in": "path",
                        "description": "ID миссии",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные результата выполнения миссии",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Файл результата",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "text": {
                                        "description": "Текстовый результат",
                                        "type": "string",
                                        "example": "Описание выполненной работы",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Результат успешно добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Mission"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "details": {
                                            "properties": {
                                                "file": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "text": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "404": {
                        "description": "Миссия не найдена"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Mission": {
                "description": "Модель задания/миссии",
                "required": [
                    "name",
                    "mission_category_id",
                    "requirement_role_id",
                    "requirement_rank_id",
                    "requirement_experience",
                    "reward_experience",
                    "reward_gold"
                ],
                "properties": {
                    "id": {
                        "description": "ID миссии",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Название миссии",
                        "type": "string",
                        "example": "Первое задание"
                    },
                    "description": {
                        "description": "Описание миссии",
                        "type": "string",
                        "example": "Выполните первое задание для получения опыта",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Иконка миссии",
                        "type": "string",
                        "example": "missions/first.png",
                        "nullable": true
                    },
                    "completion_deadline": {
                        "description": "Дедлайн выполнения миссии",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "mission_category_id": {
                        "description": "ID типа миссии",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "requirement_role_id": {
                        "description": "ID необходимой роли",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "requirement_rank_id": {
                        "description": "ID необходимого ранга",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "requirement_experience": {
                        "description": "Необходимое значение опыта",
                        "type": "integer",
                        "example": 500
                    },
                    "is_visible": {
                        "description": "Флаг видимости миссии",
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "description": "Флаг активности миссии",
                        "type": "boolean",
                        "example": true
                    },
                    "is_requirement_text": {
                        "description": "Обязательность описания результата",
                        "type": "boolean",
                        "example": false
                    },
                    "is_requirement_file": {
                        "description": "Обязательность прикрепления файла",
                        "type": "boolean",
                        "example": true
                    },
                    "reward_experience": {
                        "description": "Опыт за завершение",
                        "type": "integer",
                        "example": 100
                    },
                    "reward_gold": {
                        "description": "Золото за завершение",
                        "type": "integer",
                        "example": 50
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
...
}